fixed for borders

+++ android/src/main/java/com/shahenlibrary/Trimmer/Trimmer.java
@@ -637,7 +637,7 @@ public class Trimmer {
       mx.postScale(scaleWidth, scaleHeight);
       mx.postRotate(orientation - 360);
 
-      for (int i = startTime; i < endTime; i += step) {
+      for (int i = (int) startTime; i < (int) endTime; i += step) {
         Bitmap frame = retriever.getFrameAtTime(i * 1000);
 
         if (frame == null) {
